refactored all schema into own files
