+++ v2/server/schema/authPayload.js
@@ -1,0 +1,8 @@
+import { gql } from 'apollo-server'
+
+export default gql`
+  type AuthPayload {
+    token: String
+    user: User
+  }
+`
