+++ v2/server/schema/city.js
@@ -1,0 +1,11 @@
+import { gql } from 'apollo-server'
+
+export default gql`
+  type City {
+    id: Int!
+    name: String!
+    state: String!
+    lat: Float
+    lng: Float
+  }
+`
