+++ v2/server/schema/feedConfig.js
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server'
+
+export default gql`
+  type FeedConfig {
+    cityIds: String
+    venueTypeIds: String
+    cityDetails: [City]
+  }
+`
