+++ v2/server/schema/userProfile.js
@@ -1,0 +1,20 @@
+import { gql } from 'apollo-server'
+
+export default gql`
+  type UserProfile {
+    user: User!
+    config: UserProfileConfiguration!
+    stats: UserProfileStats!
+    recentFavoriteVenues: [Venue]
+    recentAddedVenues: [Venue]
+  }
+
+  type UserProfileConfiguration {
+    headerImageUrl: String
+  }
+
+  type UserProfileStats {
+    created: Int
+    favorited: Int
+  }
+`
